Refactor to remove Core
