+++ app/src/main/java/tv/letsrobot/controller/android/activities/SplashActivity.kt
@@ -8,8 +8,8 @@ import android.os.Bundle
 import android.widget.Toast
 import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
-import tv.letsrobot.android.api.Core
 import tv.letsrobot.android.api.interfaces.CommunicationInterface
+import tv.letsrobot.android.api.models.ServiceComponentGenerator
 import tv.letsrobot.android.api.robot.CommunicationType
 import tv.letsrobot.android.api.utils.RobotConfig
 import tv.letsrobot.controller.android.R
@@ -26,7 +26,7 @@ class SplashActivity : Activity() {
             return
         }
 
-        Core.initDependencies(this){
+        ServiceComponentGenerator.initDependencies(this){
             runOnUiThread{
                 next()
             }
